+++ R/scenarioBuilder.R
@@ -19,6 +19,7 @@
 #' 
 #' @importFrom antaresRead getAreas simOptions
 #' 
+#' @seealso \href{https://rte-antares-rpackage.github.io/antaresEditObject/articles/scenario-builder.html}{Scenario Builder vignette}
 #' @name scenario-builder
 #'
 #' @examples
@@ -65,8 +66,8 @@
 #' 
 #' # Update scenario builder
 #' 
-#' # for load serie
-#' updateScenarioBuilder(ldata = sbuilder, series = "load")
+#' # Single matrix for load serie
+#' updateScenarioBuilder(ldata = sbuilder, series = "load") # can be l instead of load
 #' 
 #' # equivalent as
 #' updateScenarioBuilder(ldata = list(l = sbuilder))
@@ -81,7 +82,7 @@
 #'   series = c("load", "hydro", "solar")
 #' )
 #' 
-#' # different input
+#' # List of matrix
 #' updateScenarioBuilder(ldata = list(
 #'   l = load_sb,
 #'   h = hydro_sb,
@@ -147,11 +148,23 @@ scenarioBuilder <- function(n_scenario,
 
 #' @title Create the correspondence data frame between the symbol and the type in scenario builder
 #' @return a `data.frame`.
-create_referential_series_type <- function(){
-
-  ref_series <- data.frame("series" = c("l", "h", "w", "s", "t", "r", "ntc", "hl"),
-                           "choices" = c("load", "hydro", "wind", "solar", "thermal", "renewables", "ntc", "hydrolevels")
-                         )
+create_scb_referential_series_type <- function(){
+  
+  series_to_write <- c("l", "h", "w", "s", "t", "r", "ntc", "hl")
+  choices <- c("load", "hydro", "wind", "solar", "thermal", "renewables", "ntc", "hydrolevels")
+  
+  # Check data consistency
+  len_series_to_write <- length(series_to_write)  
+  len_choices <- length(choices)
+  if (len_choices != len_series_to_write) {
+    stop("Inconsistent data between series and choices.\n")
+  }
+  
+  # Generate referential : w to write in scenarioBuilder, r for read only in argument
+  ref_series <- data.frame("series" = c(series_to_write, choices),
+                           "choices" = rep(choices, 2),
+                           "type" = c(rep("w",len_series_to_write), rep("r",len_choices))
+                          )
 
   return(ref_series)
 }
@@ -269,6 +282,16 @@ readScenarioBuilder <- function(ruleset = "Default Ruleset",
 #' `series = "ntc"` is only available with Antares >= 8.2.0.
 #' `series = "hl"` each value must be between 0 and 1.
 #'
+#' For a single matrix, value of series can be :
+#'  - h or hydro
+#'  - hl or hydrolevels
+#'  - l or load
+#'  - ntc
+#'  - r or renewables
+#'  - s or solar
+#'  - t or thermal
+#'  - w or wind
+#' 
 #' @export
 #' 
 #' @rdname scenario-builder
@@ -283,15 +306,17 @@ updateScenarioBuilder <- function(ldata,
   
   suppressWarnings(prevSB <- readScenarioBuilder(ruleset = ruleset, as_matrix = FALSE, opts = opts))
   
-  ref_series <- create_referential_series_type()
-  possible_series <- ref_series$series
+  ref_series <- create_scb_referential_series_type()
   
   if (!is.list(ldata)) {
     if (!is.null(series)) {
-      series <- ref_series[possible_series %in% series, "choices"]
+      if (! all(series %in% ref_series$series)) {
+        stop("Your argument series must be one of ", paste0(ref_series$series, collapse = ", "), call. = FALSE)
+      }
+      choices <- ref_series[ref_series$series %in% series, "choices"]
       if (isTRUE("ntc" %in% series) & isTRUE(opts$antaresVersion < 820))
         stop("updateScenarioBuilder: cannot use series='ntc' with Antares < 8.2.0", call. = FALSE)
-      series <- ref_series[ref_series$choices %in% series, "series"]
+      series <- ref_series[ref_series$choices %in% choices & ref_series$type == "w", "series"]
     } else {
       stop("If 'ldata' isn't a named list, you must specify which serie(s) to use!", call. = FALSE)
     }
@@ -306,8 +331,9 @@ updateScenarioBuilder <- function(ldata,
     prevSB[series] <- NULL
   } else {
     series <- names(ldata)
-    if (!all(series %in% possible_series)) {
-      stop("'ldata' must be one of ", paste0(possible_series, collapse = ", "), call. = FALSE)
+    possible_series <- ref_series[ref_series$type == "w", "series"]
+    if (! all(series %in% possible_series)) {
+      stop("Each of your list names must be in the following list : ", paste0(possible_series, collapse = ", "), call. = FALSE)
     }
     if (isTRUE("ntc" %in% series) & isTRUE(opts$antaresVersion < 820))
       stop("updateScenarioBuilder: cannot use series='ntc' with Antares < 8.2.0", call. = FALSE)